[Pallas] Pass a `kernel_info` dict to mosaic
This allows us to attach a arbitrary JSON-serializable dict to the produced custom call. This can be useful for debugging/profiling.

named_call can be used to rename a kernel in xprof or add call stack info, but it is not a good option for plumbing more structured data like block sizes, optimizations, etc.

PiperOrigin-RevId: 772683012